# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "APIs for the Github Issue Notifier Project. https://github.com/issue-notifier",
        "title": "Github Issue-Notifier API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Hemakshi Sachdev",
            "email": "sachdev.hemakshi@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8001",
    "basePath": "/",
    "paths": {
...
        "/api/v1/user/subscription/add": {
            "post": {
                "security": [
                    {
                        "Github OAuth": []
                    }
                ],
                "description": "Create new subscription for the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscription"
                ],
                "summary": "Create new subscription for the authenticated user",
                "parameters": [
                    {
                        "description": "Repository Name and list of Labels to create a new subscription",
                        "name": "subscriptions",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Subscription"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/user/subscription/remove": {
            "delete": {
                "security": [
                    {
                        "Github OAuth": []
                    }
                ],
                "description": "Deletes existing subscription for the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscription"
                ],
                "summary": "Deletes existing subscription for the authenticated user",
                "parameters": [
                    {
                        "description": "Repository Name and list of Labels which needs to be deleted from the existing subscription",
                        "name": "subscriptions",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Subscription"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/user/subscription/update": {
            "put": {
                "security": [
                    {
                        "Github OAuth": []
                    }
                ],
                "description": "Update existing subscription for the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscription"
                ],
                "summary": "Update existing subscription for the authenticated user",
                "parameters": [
                    {
                        "description": "Repository Name and list of Labels which needs to be added to the existing subscription",
                        "name": "subscriptions",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Subscription"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/user/subscription/view": {
            "get": {
                "security": [
                    {
                        "Github OAuth": []
                    }
                ],
                "description": "Get all subscriptions for the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscription"
                ],
                "summary": "Get all subscriptions for the authenticated user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Subscription"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/user/{repoName}/subscription/labels": {
            "get": {
                "security": [
                    {
                        "Github OAuth": []
                    }
                ],
                "description": "Get all subscriptions for the given `userID` and `repoName` of the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscription"
                ],
                "summary": "Get all subscriptions for the given `userID` and `repoName` of the authenticated user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repository Name for which subscription data needs to be fetched. Format `facebook/react`",
                        "name": "repoName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/models.Label"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/{repoID}/subscription/view": {
            "get": {
                "description": "Get all subscriptions for the given `repoID`",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscription"
                ],
                "summary": "Get all subscriptions for the given `repoID`",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repository ID for which subscription data needs to be fetched",
                        "name": "repoID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.UserIDLabel"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Label": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "models.Subscription": {
            "type": "object",
            "properties": {
                "labels": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Label"
                    }
                },
                "repoName": {
                    "type": "string"
                }
            }
        },
        "models.UserIDLabel": {
            "type": "object",
            "properties": {
                "label": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
...
    }
}